requirements.txt
PyYAML>=3.11
jsonrpclib>=0.1.3
